Move version functions from utils to version module
These function as version relates thus they should be in the new version
module.
